requirements.txt
opencv-python
tkinter
threading
pyautogui
Pillow


